+++ ca/django_ca/tests/base/mixins.py
@@ -53,7 +53,7 @@ from freezegun.api import FrozenDateTimeFactory, StepTickTimeFactory
 
 from django_ca import ca_settings, constants
 from django_ca.constants import ReasonFlags
-from django_ca.deprecation import RemovedInDjangoCA126Warning
+from django_ca.deprecation import RemovedInDjangoCA126Warning, RemovedInDjangoCA127Warning
 from django_ca.extensions import extension_as_text
 from django_ca.models import Certificate, CertificateAuthority, DjangoCAModel, X509CertMixin
 from django_ca.signals import (
@@ -503,6 +503,12 @@ class TestCaseMixin(TestCaseProtocol):  # pylint: disable=too-many-public-method
     def assertRemovedIn126Warning(self, msg: str) -> Iterator[None]:  # pylint: disable=invalid-name
         """Assert that a RemovedInDjangoCA126Warning is thrown."""
         with self.assertWarnsRegex(RemovedInDjangoCA126Warning, msg):
+            yield
+
+    @contextmanager
+    def assertRemovedIn127Warning(self, msg: str) -> Iterator[None]:  # pylint: disable=invalid-name
+        """Assert that a RemovedInDjangoCA127Warning is thrown."""
+        with self.assertWarnsRegex(RemovedInDjangoCA127Warning, msg):
             yield
 
     def assertRevoked(  # pylint: disable=invalid-name