+++ ca/django_ca/tests/test_settings.py
@@ -18,7 +18,6 @@ from unittest import mock
 
 from cryptography import x509
 from cryptography.hazmat.primitives import hashes
-from cryptography.hazmat.primitives.asymmetric import ec
 from cryptography.x509.oid import NameOID
 
 from django.test import TestCase